python3

+++ polytools/hexdump.py
@@ -22,7 +22,7 @@ def hexdump(buf,title="",color=6,start=0):
 	color_stop = "\033[0m"
 	out=("           %s┌"+"─"*49+"┬"+"─"*18+"┐%s\n") % (color_start,color_stop)
 	if title != "":
-		dashlen = (46-len(title))/2
+		dashlen = int((46-len(title))/2)
 		out=("           %s┌"+"─"*dashlen+"  "+title+"  "+"─"*(dashlen-(1-(len(title)%2)))+"┬"+"─"*18+"┐%s\n") % (color_start,color_stop)
 	for i in range(0,len(buf),16):
 		out+="%s0x%08x │ %s" % (color_start,i+start,color_stop)
